CAVMS-14 enable delete visitor entry

+++ protected/controllers/VisitorController.php
@@ -112,23 +112,28 @@ class VisitorController extends Controller {
      * @param integer $id the ID of the model to be deleted
      */
     public function actionDelete($id) {
-        $model = $this->loadModel($id);
-        if ($model->delete()) {
-            //throw new CHttpException(400, "This is a required field and cannot be deleted"); 
-        } else {
-            $visitorExists = Visit::model()->exists('is_deleted = 0 and visitor =' . $id . ' and (visit_status=' . VisitStatus::PREREGISTERED . ' or visit_status=' . VisitStatus::ACTIVE . ')');
-            $visitorExistsClosed = Visit::model()->exists('is_deleted = 0 and visitor =' . $id . ' and (visit_status=' . VisitStatus::CLOSED . ' or visit_status=' . VisitStatus::EXPIRED . ')');
 
-            if (!$visitorExists && !$visitorExistsClosed) {
-                
-                return false;
-            } 
+        if(Yii::app()->request->isPostRequest)
+        {
+            $model = $this->loadModel($id);
+            if ($model->delete()) {
+                //throw new CHttpException(400, "This is a required field and cannot be deleted");
+            } else {
+                $visitorExists = Visit::model()->exists('is_deleted = 0 and visitor =' . $id . ' and (visit_status=' . VisitStatus::PREREGISTERED . ' or visit_status=' . VisitStatus::ACTIVE . ')');
+                $visitorExistsClosed = Visit::model()->exists('is_deleted = 0 and visitor =' . $id . ' and (visit_status=' . VisitStatus::CLOSED . ' or visit_status=' . VisitStatus::EXPIRED . ')');
+
+                if (!$visitorExists && !$visitorExistsClosed) {
+
+                    return false;
+                }
+            }
+
+            // if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
+            if (!isset($_GET['ajax']))
+                $this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
+
         }
 
-
-        // if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
-        if (!isset($_GET['ajax']))
-            $this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
     }
 
     /**
